Refactor NTU and add support for self-paid.
